+++ crates/go_to_line/src/go_to_line.rs
@@ -85,7 +85,7 @@ impl GoToLine {
                 &editor
                     .selections
                     .last::<Point>(&editor.display_snapshot(cx)),
-                &editor.buffer().read(cx).snapshot(cx),
+                &editor.display_snapshot(cx),
             );
 
             let snapshot = active_buffer.read(cx).snapshot();
@@ -102,7 +102,7 @@ impl GoToLine {
         });
 
         let line = user_caret.line.get();
-        let column = user_caret.character.get();
+        let column = user_caret.column.get();
 
         let line_editor = cx.new(|cx| {
             let mut editor = Editor::single_line(window, cx);
@@ -623,19 +623,21 @@ mod tests {
         });
         cx.executor().advance_clock(Duration::from_millis(200));
         assert_eq!(
-            user_caret_position(1, 1),
+            user_caret_position(1, 1, 1),
             current_position(&workspace, cx),
             "Beginning of the line should be at first line, before any characters"
         );
 
+        let mut column = 1;
         for (i, c) in text.chars().enumerate() {
             let i = i as u32 + 1;
+            column += c.len_utf8() as u32;
             editor.update_in(cx, |editor, window, cx| {
                 editor.move_right(&MoveRight, window, cx)
             });
             cx.executor().advance_clock(Duration::from_millis(200));
             assert_eq!(
-                user_caret_position(1, i + 1),
+                user_caret_position(1, i + 1, column),
                 current_position(&workspace, cx),
                 "Wrong position for char '{c}' in string '{text}'",
             );
@@ -645,8 +647,9 @@ mod tests {
             editor.move_right(&MoveRight, window, cx)
         });
         cx.executor().advance_clock(Duration::from_millis(200));
+        let nchars = text.chars().count() as u32 + 1;
         assert_eq!(
-            user_caret_position(1, text.chars().count() as u32 + 1),
+            user_caret_position(1, nchars, column),
             current_position(&workspace, cx),
             "After reaching the end of the text, position should not change when moving right"
         );
@@ -700,12 +703,18 @@ mod tests {
             editor.move_to_beginning(&MoveToBeginning, window, cx)
         });
         cx.executor().advance_clock(Duration::from_millis(200));
-        assert_eq!(user_caret_position(1, 1), current_position(&workspace, cx));
+        assert_eq!(
+            user_caret_position(1, 1, 1),
+            current_position(&workspace, cx)
+        );
 
+        let mut column = 1;
         for (i, c) in text.chars().enumerate() {
             let i = i as u32 + 1;
-            let point = user_caret_position(1, i + 1);
-            go_to_point(point, user_caret_position(1, i), &workspace, cx);
+            let next_column = column + c.len_utf8() as u32;
+            let point = user_caret_position(1, i + 1, next_column);
+            go_to_point(point, user_caret_position(1, i, column), &workspace, cx);
+            column = next_column;
             cx.executor().advance_clock(Duration::from_millis(200));
             assert_eq!(
                 point,
@@ -714,15 +723,17 @@ mod tests {
             );
         }
 
+        let end_char_pos = text.chars().count() as u32 + 1;
         go_to_point(
-            user_caret_position(111, 222),
-            user_caret_position(1, text.chars().count() as u32 + 1),
+            user_caret_position(111, 222, 222),
+            user_caret_position(1, end_char_pos, column),
             &workspace,
             cx,
         );
         cx.executor().advance_clock(Duration::from_millis(200));
+        let end_char_pos = text.chars().count() as u32 + 1;
         assert_eq!(
-            user_caret_position(1, text.chars().count() as u32 + 1),
+            user_caret_position(1, end_char_pos, column),
             current_position(&workspace, cx),
             "When going into too large point, should go to the end of the text"
         );
@@ -744,10 +755,11 @@ mod tests {
         })
     }
 
-    fn user_caret_position(line: u32, character: u32) -> UserCaretPosition {
+    fn user_caret_position(line: u32, character: u32, column: u32) -> UserCaretPosition {
         UserCaretPosition {
             line: NonZeroU32::new(line).unwrap(),
             character: NonZeroU32::new(character).unwrap(),
+            column: NonZeroU32::new(column).unwrap(),
         }
     }
 
@@ -767,7 +779,7 @@ mod tests {
                 }),
                 format!(
                     "{}:{}",
-                    expected_placeholder.line, expected_placeholder.character
+                    expected_placeholder.line, expected_placeholder.column
                 )
             );
         });